HARMONY-1891: Move copy handler to common util file
